feat: move tests into their own files
