+++ auth-service/tests/api/routes/logout.rs
@@ -1,0 +1,11 @@
+use crate::helpers::TestApp;
+
+#[tokio::test]
+async fn logout_returns_200() {
+    let app = TestApp::new().await;
+
+    let response = app.post_logout().await;
+
+    assert_eq!(response.status().as_u16(), 200);
+}
+
