+++ auth-service/tests/api/routes/root.rs
@@ -1,0 +1,16 @@
+use crate::helpers::TestApp;
+
+// Tokio's test macro is used to run the test in an async environment
+#[tokio::test]
+async fn root_returns_auth_ui() {
+    let app = TestApp::new().await;
+
+    let response = app.get_root().await;
+
+    assert_eq!(response.status().as_u16(), 200);
+    assert_eq!(response.headers().get("content-type").unwrap(), "text/html");
+}
+
+// TODO: Implement tests for all other routes (signup, login, logout, verify-2fa, and verify-token)
+// For now, simply assert that each route returns a 200 HTTP status code.
+
